feat: refactor useChatTour hook and update import paths for consistency
