+++ src/components/nav-main.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/sidebar";
 import { cn } from "@/lib/utils";
 import { Link, useLocation } from "react-router";
-import useChatTour from "@/sections/chat/hooks/use-chat-tour";
+import useChatTour from "@/hooks/use-chat-tour";
 
 export function NavMain({
   items,
